Tidy up nym-wallet-types network config
