+++ nym-wallet/nym-wallet-types/src/network.rs
@@ -7,6 +7,9 @@ use serde::{Deserialize, Serialize};
 use std::{fmt, ops::Not, str::FromStr};
 use strum::EnumIter;
 
+mod qa;
+mod sandbox;
+
 #[allow(clippy::upper_case_acronyms)]
 #[cfg_attr(feature = "generate-ts", derive(ts_rs::TS))]
 #[cfg_attr(
@@ -91,139 +94,4 @@ impl FromStr for Network {
 
 fn parse_optional_str(raw: &str) -> Option<String> {
     raw.is_empty().not().then(|| raw.into())
-}
-
-mod sandbox {
-    use nym_network_defaults::{ChainDetails, DenomDetails, NymContracts, ValidatorDetails};
-
-    use super::parse_optional_str;
-
-    pub(crate) const BECH32_PREFIX: &str = "nymt";
-
-    pub(crate) const MIX_DENOM: DenomDetails = DenomDetails::new("unymt", "nymt", 6);
-    pub(crate) const STAKE_DENOM: DenomDetails = DenomDetails::new("unyxt", "nyxt", 6);
-
-    pub(crate) const MIXNET_CONTRACT_ADDRESS: &str = "nymt1ghd753shjuwexxywmgs4xz7x2q732vcnstz02j";
-    pub(crate) const VESTING_CONTRACT_ADDRESS: &str = "nymt14ejqjyq8um4p3xfqj74yld5waqljf88fn549lh";
-    pub(crate) const BANDWIDTH_CLAIM_CONTRACT_ADDRESS: &str =
-        "nymt17p9rzwnnfxcjp32un9ug7yhhzgtkhvl9f8xzkv";
-    pub(crate) const COCONUT_BANDWIDTH_CONTRACT_ADDRESS: &str =
-        "nymt1nz0r0au8aj6dc00wmm3ufy4g4k86rjzlgq608r";
-    pub(crate) const GROUP_CONTRACT_ADDRESS: &str = "nymt1k8re7jwz6rnnwrktnejdwkwnncte7ek7kk6fvg";
-    pub(crate) const MULTISIG_CONTRACT_ADDRESS: &str =
-        "nymt1k8re7jwz6rnnwrktnejdwkwnncte7ek7kk6fvg";
-    pub(crate) const COCONUT_DKG_CONTRACT_ADDRESS: &str =
-        "nymt1k8re7jwz6rnnwrktnejdwkwnncte7ek7kk6fvg";
-    pub(crate) const SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS: &str = "";
-    pub(crate) const NAME_SERVICE_CONTRACT_ADDRESS: &str = "";
-    pub(crate) const _ETH_CONTRACT_ADDRESS: [u8; 20] =
-        hex_literal::hex!("8e0DcFF7F3085235C32E845f3667aEB3f1e83133");
-    pub(crate) const _ETH_ERC20_CONTRACT_ADDRESS: [u8; 20] =
-        hex_literal::hex!("E8883BAeF3869e14E4823F46662e81D4F7d2A81F");
-    //pub(crate) const REWARDING_VALIDATOR_ADDRESS: &str =
-    //"nymt1jh0s6qu6tuw9ut438836mmn7f3f2wencrnmdj4";
-
-    //pub(crate) const STATISTICS_SERVICE_DOMAIN_ADDRESS: &str = "http://0.0.0.0";
-    pub(crate) fn validators() -> Vec<ValidatorDetails> {
-        vec![ValidatorDetails::new(
-            "https://sandbox-validator.nymtech.net",
-            Some("https://sandbox-validator.nymtech.net/api"),
-        )]
-    }
-
-    pub(crate) fn network_details() -> nym_network_defaults::NymNetworkDetails {
-        nym_network_defaults::NymNetworkDetails {
-            chain_details: ChainDetails {
-                bech32_account_prefix: BECH32_PREFIX.to_string(),
-                mix_denom: MIX_DENOM.into(),
-                stake_denom: STAKE_DENOM.into(),
-            },
-            endpoints: validators(),
-            contracts: NymContracts {
-                mixnet_contract_address: parse_optional_str(MIXNET_CONTRACT_ADDRESS),
-                vesting_contract_address: parse_optional_str(VESTING_CONTRACT_ADDRESS),
-                bandwidth_claim_contract_address: parse_optional_str(
-                    BANDWIDTH_CLAIM_CONTRACT_ADDRESS,
-                ),
-                coconut_bandwidth_contract_address: parse_optional_str(
-                    COCONUT_BANDWIDTH_CONTRACT_ADDRESS,
-                ),
-                group_contract_address: parse_optional_str(GROUP_CONTRACT_ADDRESS),
-                multisig_contract_address: parse_optional_str(MULTISIG_CONTRACT_ADDRESS),
-                coconut_dkg_contract_address: parse_optional_str(COCONUT_DKG_CONTRACT_ADDRESS),
-                service_provider_directory_contract_address: parse_optional_str(
-                    SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS,
-                ),
-                name_service_contract_address: parse_optional_str(NAME_SERVICE_CONTRACT_ADDRESS),
-            },
-        }
-    }
-}
-
-mod qa {
-    use nym_network_defaults::{ChainDetails, DenomDetails, NymContracts, ValidatorDetails};
-
-    use super::parse_optional_str;
-
-    pub(crate) const BECH32_PREFIX: &str = "n";
-
-    pub(crate) const MIX_DENOM: DenomDetails = DenomDetails::new("unym", "nym", 6);
-    pub(crate) const STAKE_DENOM: DenomDetails = DenomDetails::new("unyx", "nyx", 6);
-
-    pub(crate) const MIXNET_CONTRACT_ADDRESS: &str =
-        "n14hj2tavq8fpesdwxxcu44rty3hh90vhujrvcmstl4zr3txmfvw9sjyvg3g";
-    pub(crate) const VESTING_CONTRACT_ADDRESS: &str =
-        "n1nc5tatafv6eyq7llkr2gv50ff9e22mnf70qgjlv737ktmt4eswrq73f2nw";
-    pub(crate) const BANDWIDTH_CLAIM_CONTRACT_ADDRESS: &str =
-        "n19lc9u84cz0yz3fww5283nucc9yvr8gsjmgeul0";
-    pub(crate) const COCONUT_BANDWIDTH_CONTRACT_ADDRESS: &str =
-        "n1ghd753shjuwexxywmgs4xz7x2q732vcn7ty4yw";
-    pub(crate) const GROUP_CONTRACT_ADDRESS: &str = "n17p9rzwnnfxcjp32un9ug7yhhzgtkhvl988qccs";
-    pub(crate) const MULTISIG_CONTRACT_ADDRESS: &str = "n17p9rzwnnfxcjp32un9ug7yhhzgtkhvl988qccs";
-    pub(crate) const COCONUT_DKG_CONTRACT_ADDRESS: &str =
-        "n17p9rzwnnfxcjp32un9ug7yhhzgtkhvl988qccs";
-    pub(crate) const SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS: &str =
-        "n1ryt076cufyddallg5x0gz3qjz0pd3wg0m4cwkg9njhmlnp6u88qq6nczgj";
-    pub(crate) const NAME_SERVICE_CONTRACT_ADDRESS: &str = "";
-    pub(crate) const _ETH_CONTRACT_ADDRESS: [u8; 20] =
-        hex_literal::hex!("0000000000000000000000000000000000000000");
-    pub(crate) const _ETH_ERC20_CONTRACT_ADDRESS: [u8; 20] =
-        hex_literal::hex!("0000000000000000000000000000000000000000");
-    //pub(crate) const REWARDING_VALIDATOR_ADDRESS: &str = "n17dcjwykjj9ydd8p9hk65uf2flrhhlfklju8flp";
-
-    //pub(crate) const STATISTICS_SERVICE_DOMAIN_ADDRESS: &str = "http://0.0.0.0";
-    pub(crate) fn validators() -> Vec<ValidatorDetails> {
-        vec![ValidatorDetails::new(
-            "https://qwerty-validator.qa.nymte.ch/",
-            Some("https://qwerty-validator-api.qa.nymte.ch/api"),
-        )]
-    }
-
-    pub(crate) fn network_details() -> nym_network_defaults::NymNetworkDetails {
-        nym_network_defaults::NymNetworkDetails {
-            chain_details: ChainDetails {
-                bech32_account_prefix: BECH32_PREFIX.to_string(),
-                mix_denom: MIX_DENOM.into(),
-                stake_denom: STAKE_DENOM.into(),
-            },
-            endpoints: validators(),
-            contracts: NymContracts {
-                mixnet_contract_address: parse_optional_str(MIXNET_CONTRACT_ADDRESS),
-                vesting_contract_address: parse_optional_str(VESTING_CONTRACT_ADDRESS),
-                bandwidth_claim_contract_address: parse_optional_str(
-                    BANDWIDTH_CLAIM_CONTRACT_ADDRESS,
-                ),
-                coconut_bandwidth_contract_address: parse_optional_str(
-                    COCONUT_BANDWIDTH_CONTRACT_ADDRESS,
-                ),
-                group_contract_address: parse_optional_str(GROUP_CONTRACT_ADDRESS),
-                multisig_contract_address: parse_optional_str(MULTISIG_CONTRACT_ADDRESS),
-                coconut_dkg_contract_address: parse_optional_str(COCONUT_DKG_CONTRACT_ADDRESS),
-                service_provider_directory_contract_address: parse_optional_str(
-                    SERVICE_PROVIDER_DIRECTORY_CONTRACT_ADDRESS,
-                ),
-                name_service_contract_address: parse_optional_str(NAME_SERVICE_CONTRACT_ADDRESS),
-            },
-        }
-    }
 }
